ci: revert adding acceptance concurrency group also to release
As the concurrency groups are defined separately from each other
they don't cancel the remaining workflow as a whole.
Hence resulting in undesired behaviour as the release could happen,
while the acceptance tests didn't executed.
Normally this should be noticed by the cancel that happened on acceptance,
but that's explicitly ignored as a work-around for the skipping of the
devcontainer.

This reverts commit fd23faeae5c40f54739dd6d82b0afccd35796aa1.